Create file createCalendar and add code to create calendar from main.ts
